Move caretToEnd logic out of focus handling
