+++ ts/editor/changeTimer.ts
@@ -6,16 +6,12 @@ import type { EditingArea } from ".";
 import { getCurrentField } from ".";
 import { bridgeCommand } from "./lib";
 import { getNoteId } from "./noteId";
-import { updateButtonState } from "./toolbar";
 
 let changeTimer: number | null = null;
 
 export function triggerChangeTimer(currentField: EditingArea): void {
     clearChangeTimer();
-    changeTimer = setTimeout(function () {
-        updateButtonState();
-        saveField(currentField, "key");
-    }, 600);
+    changeTimer = setTimeout(() => saveField(currentField, "key"), 600);
 }
 
 function clearChangeTimer(): void {
